+++ packages/connectivity/src/scp-cf/destination/destination-service.spec.ts
@@ -2,6 +2,7 @@ import nock from 'nock';
 import * as jwt123 from 'jsonwebtoken';
 import axios, { AxiosRequestConfig } from 'axios';
 import { HttpsProxyAgent } from 'https-proxy-agent';
+import { createLogger } from '@sap-cloud-sdk/util';
 import { destinationServiceUri } from '../../../../../test-resources/test/test-util/environment-mocks';
 import { privateKey } from '../../../../../test-resources/test/test-util/keys';
 import { defaultResilienceBTPServices } from '../resilience-options';
@@ -50,6 +51,11 @@ const oauth2SamlBearerDestination = {
   tokenServicePassword: 'password'
 };
 
+const brokenDestination = {
+  Name: 'BrokenDestination',
+  URL: undefined
+};
+
 describe('destination service', () => {
   describe('fetchInstanceDestinations', () => {
     it('fetches instance destinations and returns them as Destination array', async () => {
@@ -93,6 +99,36 @@ describe('destination service', () => {
       expected.forEach((e, index) => {
         expect(instanceDestinations[index]).toMatchObject(e);
       });
+    });
+
+    it('only returns valid destinations - instance destinations', async () => {
+      const response = [
+        basicDestination,
+        oauth2SamlBearerDestination,
+        brokenDestination
+      ];
+
+      nock(destinationServiceUri, {
+        reqheaders: {
+          authorization: `Bearer ${jwt}`
+        }
+      })
+        .get('/destination-configuration/v1/instanceDestinations')
+        .reply(200, response);
+
+      const logger = createLogger({
+        package: 'connectivity',
+        messageContext: 'destination-service'
+      });
+      const debugSpy = jest.spyOn(logger, 'debug');
+      const instanceDestinations: Destination[] =
+        await fetchInstanceDestinations(destinationServiceUri, jwt, {
+          enableCircuitBreaker: false
+        });
+      expect(instanceDestinations.length).toBe(2);
+      expect(debugSpy).toHaveBeenCalledWith(
+        'Parsing of destination with name "BrokenDestination" failed - skip this destination in parsing.'
+      );
     });
 
     it('returns 400 for an invalid JWT', async () => {
@@ -173,6 +209,36 @@ describe('destination service', () => {
       expected.forEach((e, index) => {
         expect(subaccountDestinations[index]).toMatchObject(e);
       });
+    });
+
+    it('only returns valid destinations - subaccount destinations', async () => {
+      const response = [
+        basicDestination,
+        oauth2SamlBearerDestination,
+        brokenDestination
+      ];
+
+      nock(destinationServiceUri, {
+        reqheaders: {
+          authorization: `Bearer ${jwt}`
+        }
+      })
+        .get('/destination-configuration/v1/subaccountDestinations')
+        .reply(200, response);
+
+      const logger = createLogger({
+        package: 'connectivity',
+        messageContext: 'destination-service'
+      });
+      const debugSpy = jest.spyOn(logger, 'debug');
+      const subaccountDestinations: Destination[] =
+        await fetchSubaccountDestinations(destinationServiceUri, jwt, {
+          enableCircuitBreaker: false
+        });
+      expect(subaccountDestinations.length).toBe(2);
+      expect(debugSpy).toHaveBeenCalledWith(
+        'Parsing of destination with name "BrokenDestination" failed - skip this destination in parsing.'
+      );
     });
 
     it('returns 400 for an invalid JWT', async () => {