+++ test-packages/test-services-odata-v4/test-service/BatchRequest.d.ts
@@ -6,6 +6,7 @@ import {
   GetByKeyRequestBuilder,
   ODataBatchRequestBuilder,
   UpdateRequestBuilder,
+  FunctionImportRequestBuilder,
   BatchChangeSet
 } from '@sap-cloud-sdk/odata-v4';
 import {
@@ -22,7 +23,19 @@ import {
   TestEntityCircularLinkParent,
   TestEntityCircularLinkChild,
   TestEntityEndsWith,
-  TestEntityEndsWithSomethingElse
+  TestEntityEndsWithSomethingElse,
+  TestFunctionImportEdmReturnTypeParameters,
+  TestFunctionImportEdmReturnTypeCollectionParameters,
+  TestFunctionImportNullableTestParameters,
+  TestFunctionImportEntityReturnTypeParameters,
+  TestFunctionImportEntityReturnTypeCollectionParameters,
+  TestFunctionImportSharedEntityReturnTypeParameters,
+  TestFunctionImportSharedEntityReturnTypeCollectionParameters,
+  TestFunctionImportComplexReturnTypeParameters,
+  TestFunctionImportComplexReturnTypeCollectionParameters,
+  TestFunctionImportMultipleParamsParameters,
+  TestFunctionImportWithDifferentNameParameters,
+  TestComplexType
 } from './index';
 /**
  * Batch builder for operations supported on the Test Service.
@@ -140,6 +153,67 @@ export declare type ReadTestServiceRequestBuilder<
   | GetByKeyRequestBuilder<
       TestEntityEndsWithSomethingElse<DeSerializersT>,
       DeSerializersT
+    >
+  | FunctionImportRequestBuilder<
+      DeSerializersT,
+      TestFunctionImportEdmReturnTypeParameters<DeSerializersT>,
+      boolean
+    >
+  | FunctionImportRequestBuilder<
+      DeSerializersT,
+      TestFunctionImportEdmReturnTypeCollectionParameters<DeSerializersT>,
+      string[]
+    >
+  | FunctionImportRequestBuilder<
+      DeSerializersT,
+      TestFunctionImportNullableTestParameters<DeSerializersT>,
+      string[] | null
+    >
+  | FunctionImportRequestBuilder<
+      DeSerializersT,
+      TestFunctionImportEntityReturnTypeParameters<DeSerializersT>,
+      TestEntity
+    >
+  | FunctionImportRequestBuilder<
+      DeSerializersT,
+      TestFunctionImportEntityReturnTypeCollectionParameters<DeSerializersT>,
+      TestEntity[]
+    >
+  | Omit<
+      FunctionImportRequestBuilder<
+        DeSerializersT,
+        TestFunctionImportSharedEntityReturnTypeParameters<DeSerializersT>,
+        never
+      >,
+      'execute'
+    >
+  | Omit<
+      FunctionImportRequestBuilder<
+        DeSerializersT,
+        TestFunctionImportSharedEntityReturnTypeCollectionParameters<DeSerializersT>,
+        never
+      >,
+      'execute'
+    >
+  | FunctionImportRequestBuilder<
+      DeSerializersT,
+      TestFunctionImportComplexReturnTypeParameters<DeSerializersT>,
+      TestComplexType
+    >
+  | FunctionImportRequestBuilder<
+      DeSerializersT,
+      TestFunctionImportComplexReturnTypeCollectionParameters<DeSerializersT>,
+      TestComplexType[]
+    >
+  | FunctionImportRequestBuilder<
+      DeSerializersT,
+      TestFunctionImportMultipleParamsParameters<DeSerializersT>,
+      boolean | null
+    >
+  | FunctionImportRequestBuilder<
+      DeSerializersT,
+      TestFunctionImportWithDifferentNameParameters<DeSerializersT>,
+      undefined
     >;
 export declare type WriteTestServiceRequestBuilder<
   DeSerializersT extends DeSerializers