+++ packages/eslint-config/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @sap-cloud-sdk/eslint-config
 
+## 2.7.1
+
+## 2.7.0
+
 ## 2.6.0
 
 ## 2.5.0