+++ packages/generator-common/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-cloud-sdk/generator-common
 
+## 2.7.1
+
+### Patch Changes
+
+- @sap-cloud-sdk/util@2.7.1
+
+## 2.7.0
+
+### Patch Changes
+
+- @sap-cloud-sdk/util@2.7.0
+
 ## 2.6.0
 
 ### Patch Changes