+++ packages/test-util/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @sap-cloud-sdk/test-util
 
+## 2.7.1
+
+### Patch Changes
+
+- @sap-cloud-sdk/connectivity@2.7.1
+- @sap-cloud-sdk/util@2.7.1
+
+## 2.7.0
+
+### Patch Changes
+
+- Updated dependencies [3bff42e1]
+- Updated dependencies [010a46fa]
+  - @sap-cloud-sdk/connectivity@2.7.0
+  - @sap-cloud-sdk/util@2.7.0
+
 ## 2.6.0
 
 ### Patch Changes