+++ test-packages/test-services-odata-v2/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @sap-cloud-sdk/test-services-odata-v2
 
+## 2.7.1
+
+## 2.7.0
+
 ## 2.6.0
 
 ### Patch Changes